add overwrite to download_raster_entry

+++ rgdc/rgdc/utils.py
@@ -41,13 +41,14 @@ def datetime_to_str(value: object):
     return value
 
 
-def download_checksum_file_to_path(file: Dict, path: Path):
+def download_checksum_file_to_path(file: Dict, path: Path, overwrite: bool):
     """
     Download a RGD ChecksumFile to a given path.
 
     Args:
         file: A RGD ChecksumFile serialized as a Dict.
         path: The root path to download this file to.
+        overwrite: if True, replace files existing on disk
 
     Returns:
         The path on disk the file was downloaded to.
@@ -72,8 +73,9 @@ def download_checksum_file_to_path(file: Dict, path: Path):
 
     # Download contents to file
     file_path = parent_path / filename
-    with open(file_path, 'wb') as open_file_path:
-        for chunk in iterate_response_bytes(file_download_url):
-            open_file_path.write(chunk)
+    if not file_path.is_file() or overwrite:
+        with open(file_path, 'wb') as open_file_path:
+            for chunk in iterate_response_bytes(file_download_url):
+                open_file_path.write(chunk)
 
     return file_path
